Changes to make disk video work at resolutions up to 46340 X 46340.  Up to 65535 X 65535 should work, but we are limited by the file access routines built into gcc that uses a signed long value for determining the position in a file.

+++ branches/ssg/sdl/common/realdos.c
@@ -1681,8 +1681,8 @@ int load_fractint_cfg(int options)
       vident->name[25] = vident->comment[25] = 0;
       vident->keynum      = keynum;
       vident->dotmode     = dotmode;
-      vident->xdots       = (short)xdots;
-      vident->ydots       = (short)ydots;
+      vident->xdots       = (unsigned int)xdots;
+      vident->ydots       = (unsigned int)ydots;
       vident->colors      = colors;
       ++vident;
       ++vidtbllen;
